Add options to facilitate research
Added a commented out Println() in the Get function, which can be
uncommented to output the JSON gotten from the HTTP response body.

FakeGet() funtion added to aid with testing without depending on the
ability to access and edit actual Salesforce records. It will Unmarshal
the passed local JSON file and proceed as if the JSON were gotten from
an HTTP request.